test: split scheduler-app checks — E008 (app absent) to core, W003 to pg_cron
